backend-GetAllUsersOfTheSystem: Moved the classes to the corresponding packages
