+++ todo-backend/src/main/java/com/umss/todo/controller/UserRestController.java
@@ -1,4 +1,4 @@
-package com.umss.todo;
+package com.umss.todo.controller;
 
 import java.util.List;
 
@@ -8,23 +8,27 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.umss.todo.User;
+import com.umss.todo.reposiroty.SqlConnector;
+import com.umss.todo.service.UserService;
+
 @RestController
 @RequestMapping("/users")
 public class UserRestController {
 	
 	// FEATURE: backend-GetAllUsersOfTheSystem	
-	private SqlConnector sqlConnector;
+	private UserService userService;
 	
 	
 	@Autowired
-	public UserRestController(@Qualifier("mysql") SqlConnector sqlConnector) {
-		this.sqlConnector = sqlConnector;
+	public UserRestController(UserService userService) {
+		this.userService = userService;
 	}
 	
 	
 	// http://localhost:8080/users
 	@GetMapping
 	public List<User> getAllSystemUsers() {
-		return sqlConnector.getAllUsers();
+		return userService.getAllUsers();
 	}
 }
